+++ lib/keys/filters.js
@@ -1,4 +1,6 @@
 'use strict'
+const sortAdd = require('./sort/filters').add
+
 exports.create = function (target) {
   if (!target._filters) {
     target._filters = {}
@@ -19,56 +21,12 @@ exports.add = function addFilter (target, key, copy, sIndex) {
   const filtered = type && filters[type]
   if (filtered) {
     if (sIndex !== void 0) {
-      let keys = target._keys
-      if (filtered[sIndex] !== keys[sIndex]) {
-        let indexOfNext = -1
-        let i = sIndex + 1
-        // can be greatly optmized
-        while (indexOfNext === -1 && i < keys.length) {
-          indexOfNext = indexOf(filtered, keys[i])
-          i++
-        }
-        if (indexOfNext !== -1) {
-          filtered.splice(indexOfNext, 0, key)
-        } else {
-          filtered.push(key)
-        }
-      } else {
-        filtered.push(key)
-      }
+      sortAdd(target, key, filtered, sIndex)
     } else {
       filtered.push(key)
     }
     console.log('need to update instances as well! (add filter)')
   }
-}
-
-exports.resort = function resort (target, keys, field, left, right) {
-  console.log('FILTERS - needs optmization')
-  const filters = target._filters
-  for (let type in filters) {
-    filters[type] = resortKeys(filters[type], keys)
-    console.log('need to update instances as well!')
-  }
-}
-
-function resortKeys (filter, keys) {
-  // use quick for this
-  // this is the worsed do it with swaps please
-  var last = -1
-  const arr = []
-  for (let i = 0, len = filter.length; i < len; i++) {
-    let index = indexOf(keys, filter[i])
-    if (index !== -1) {
-      if (index >= last) {
-        arr.push(filter[i])
-      } else {
-        arr.unshift(filter[i])
-      }
-      last = index
-    }
-  }
-  return arr
 }
 
 exports.remove = function removeFilter (target, key, copy) {
@@ -105,12 +63,4 @@ function copyFilters (filters) {
     newFilters[type] = filters[type].concat()
   }
   return newFilters
-}
-
-function indexOf (arr, key) {
-  for (let i = 0, len = arr.length; i < len; i++) {
-    if (arr[i] === key) {
-      return i
-    }
-  }
 }
