+++ lib/keys/sort/index.js
@@ -1,14 +1,16 @@
 'use strict'
 const quickSort = require('./quick')
 const order = require('./order').compute
-const resortFilters = require('../filters').resort
+const resortFilters = require('./filters').resort
 
+// resort option that makes a new orders array
 module.exports = exports = function (target, keys, field, left, right) {
+  console.log('resort -- needs to update instances as well (SORT/index)')
+  const orders = keys._ || (keys._ = [])
   if (!right) { right = keys.length - 1 }
   if (!left) { left = 0 }
-  const orders = keys._ || (keys._ = [])
-  if (!right) {
-    if (keys.length) {
+  if (right < 1) {
+    if (right === 0) {
       orders[0] = order(target[keys[0]], field)
     }
   } else {
